+++ src/save.jl
@@ -53,8 +53,8 @@ function filesstorage_settings(figname)
 end
 #fig2
 
-floattostring(vec::Array{Float32,1}) = [floattostring(vec[i]) for i in 1:length(vec)]
-floattostring(i::Float32) = "$(Int(i))"
+floattostring(vec::Array{Float64,1}) = [floattostring(vec[i]) for i in 1:length(vec)]
+floattostring(i::Float64) = "$(Int(i))"
 function writecsv(vec, p, name, taglist)
     karray = collect(Iterators.flatten(
     [collect(Iterators.flatten([append!([],vec[1][1])
